chore(web): trim comment bloat and drop dead code
Comments across web/ had drifted into narrating their own development
history — task-report references (to files that are gitignored and so
don't exist for anyone else), "the old code did X", "Fix round 1",
per-PR archaeology, and paragraph-long restatements of what the code
below already says.

Rewrote the worst offenders to keep the load-bearing half — the browser
quirks, the RHF/zod pitfalls, the fail-open/fail-closed rules — and drop
the rest. ~570 comment lines out, no behaviour change.

Dead code removed while in there:
- components/admin/LibraryPlaylistsTab.tsx — orphaned when LibraryPanel
  was split into per-tab components; nothing rendered it. Its one live
  export, PlaylistSummary, moved to library/types.ts.
- FormUpdaterOrErrors (settings/shared.tsx), PROMPT_MIN
  (personas/constants.ts), SwitchRow (playlist-builder/bits.tsx),
  PersonaPicker (shows/ShowPickers.tsx) — all unreferenced.

Also fixed a stacked JSDoc block in settings/shared.tsx where
settingsFieldAria's doc comment sat above SettingsFieldError instead.
