+++ web/components/admin/skills/SkillEditModal.tsx
@@ -111,14 +111,10 @@ interface SkillFileResponse {
 
 const COOLDOWN_PRESETS = ['15m', '25m', '45m', '1h', '6h'];
 
-// The RHF-bound shape of the SKILL.md fields (what `useZodForm`'s schema
-// validates). `name` is create-only — the edit-mode schema (skillFileSchema)
-// has no such field, so it just rides along unused there. `config` (the
-// skill's own declared knobs) is deliberately NOT here: those are runtime
-// data from the skill's own tool.mjs, validated separately by
-// skills/config-fields.ts on the controller, so they stay their own
-// useState below rather than joining the shared schema — same reasoning as
-// the schema file's own header comment.
+// The RHF-bound shape of the SKILL.md fields. `name` is create-only and rides
+// along unused in edit mode. `config` (the skill's own declared knobs) stays
+// out: it's runtime data from the skill's tool.mjs, validated separately by
+// the controller's skills/config-fields.ts, so it keeps its own useState.
 interface SkillFormValues {
   name?: string;
   label: string;
@@ -210,15 +206,11 @@ export default function SkillEditModal({ mode, skill, personas, tagSuggestions, 
   const [confirmDelete, setConfirmDelete] = useState(false);  // delete confirm dialog
   const [defaults, setDefaults] = useState<SkillDefaults | null>(null); // built-in shipped defaults
 
-  // The same schema the controller runs (controller/src/schemas/skill.ts via
-  // the generated mirror) — so a bad cooldown is caught at the input instead of
-  // coming back as a 400 after the operator hits Save. Declared as the widened
-  // ZodType rather than the literal create/edit union: several of its fields
-  // (label/cooldown/window/requiresKey) are z.preprocess-wrapped, whose
-  // z.input is `unknown`, so the union's input type collapses the same way a
-  // factory schema's does (see FestivalsSection/MoodsPanel) — worked around
-  // the same way, with one cast on `control` below instead of fighting the
-  // union at every call site.
+  // The same schema the controller runs, so a bad cooldown is caught at the
+  // input rather than coming back as a 400. Declared as the widened ZodType
+  // rather than the create/edit union: several fields are z.preprocess-wrapped,
+  // whose z.input is `unknown`, so the union's input type collapses — handled
+  // with one cast on `control` below instead of at every call site.
   const schema: z.ZodType<FieldValues, FieldValues> =
     mode === 'create' ? skillCreateSchema : skillFileSchema(custom);
 
@@ -232,13 +224,10 @@ export default function SkillEditModal({ mode, skill, personas, tagSuggestions, 
   const control = form.control as unknown as Control<SkillFormValues>;
   const uid = useId();
 
-  // `custom` can flip after mount (edit mode's initial guess comes from the
-  // skills-list row; the file GET below is the source of truth), which swaps
-  // `schema` to a different singleton. react-hook-form picks up a changed
-  // resolver on the next render, but it doesn't retroactively re-run it
-  // against already-computed error state — only the next trigger does. Same
-  // pattern as MoodsPanel's schedule/weather schema, which also depends on
-  // state resolved after mount.
+  // `custom` can flip after mount (the file GET below corrects the list row's
+  // guess), swapping `schema`. RHF picks up the new resolver on the next render
+  // but won't re-run it against already-computed error state — same as
+  // MoodsPanel's schedule/weather schema.
   useEffect(() => {
     void form.trigger();
     // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -308,15 +297,11 @@ export default function SkillEditModal({ mode, skill, personas, tagSuggestions, 
   const dirty = loaded && (form.formState.isDirty || configDirty || assignDirty);
 
   const canSave = loaded && form.formState.isValid && !busy;
-  // Every field has its own inline error slot now except `requiresKey`
-  // (name/label/cooldown/brief via TextField/TextareaField's built-in
-  // FieldError; context/tags/window via the hand-rolled Controller blocks
-  // below) — excluded here so a message doesn't render TWICE, once under the
-  // field and again in this generic banner. `requiresKey` is the one field
-  // with NO rendered control at all (a hidden passthrough — see its
-  // declaration above), so a disk-authored value that isn't UPPER_SNAKE_CASE
-  // has nowhere else to surface. The footer shows the first such issue so a
-  // gated Save always says why.
+  // Every field has its own inline error slot, so listing them here too would
+  // render each message twice. `requiresKey` is the exception: it's a hidden
+  // passthrough with no rendered control, so a disk-authored value that isn't
+  // UPPER_SNAKE_CASE has nowhere else to surface, and a gated Save would
+  // otherwise never say why.
   const FIELDS_WITH_INLINE_ERRORS = ['name', 'label', 'cooldown', 'context', 'tags', 'window', 'brief'];
   const blockingIssue = (() => {
     const entry = Object.entries(form.formState.errors).find(
@@ -339,15 +324,13 @@ export default function SkillEditModal({ mode, skill, personas, tagSuggestions, 
     setBusy(true);
     try {
       // `requiresKey` (and, for a custom skill, `window`) ride along in `values`
-      // whenever the schema in force declares them — a built-in edit's schema
-      // (builtinSkillFileSchema) doesn't, so zod has already stripped them from
-      // the parsed output, same as the old `...(custom ? {…} : {})` spread.
+      // only when the schema in force declares them — a built-in edit's schema
+      // doesn't, so zod has already stripped them from the parsed output.
       const body: Record<string, unknown> = { ...values };
       // Always sent when the skill declares knobs, so clearing a field clears
-      // the frontmatter line. Omitted entirely for a skill with none, which the
-      // controller reads as "leave whatever is on disk". `config` is read off
-      // the raw body server-side (routes/dj.ts's resolveConfig), never off the
-      // parsed schema output, so it travels outside `values` here too.
+      // the frontmatter line; omitted for a skill with none, which the
+      // controller reads as "leave whatever is on disk". Read off the raw body
+      // server-side, so it travels outside `values` here too.
       if (configFields.length) {
         body.config = Object.fromEntries(
           configFields.map(f => [f.key, (config[f.key] || '').trim()]),
@@ -375,9 +358,8 @@ export default function SkillEditModal({ mode, skill, personas, tagSuggestions, 
       };
       if (!r.ok) {
         // A server-side name rule (reserved slug, slug already on disk) comes
-        // back as fieldErrors.name — typing a different slug is the way out,
-        // so it lands on the slug input rather than only flashing past in a
-        // toast.
+        // back as fieldErrors.name. Typing a different slug is the way out, so
+        // land it on the input rather than only in a toast.
         applyServerFieldErrors(form, j.fieldErrors);
         throw new Error(j.error || `failed (${r.status})`);
       }
