+++ web/components/admin/skills/contextFields.ts
@@ -1,8 +1,6 @@
 // Shared "right now" context-field vocabulary for the Skills admin (#471).
 //
-// The vocabulary itself comes from the shared skill schema
-// (controller/src/schemas/skill.ts, mirrored into lib/schemas.generated.ts) —
-// this file used to carry a hand-maintained second copy of it.
+// The vocabulary itself comes from the mirrored skill schema.
 import { CONTEXT_FIELDS, type ContextField } from '@/lib/schemas.generated';
 
 // Human labels for the chips. Typed as an EXHAUSTIVE record, so adding a field
