+++ web/lib/site.ts
@@ -32,11 +32,10 @@ function isOfficialHost(url: string): boolean {
 // opted into indexing everything with SUBWAVE_INDEX_ALL=1.
 //
 // Every install serves the same marketing/docs/news pages as getsubwave.com
-// (one shared image), and each used to self-assert canonical over them — so
-// Google folds the duplicates into ONE winner per cluster, and a public
-// self-hosted station could win a dispatch or setup guide away from the
-// official site. Non-official installs therefore donate those canonicals
-// back (and drop the pages from their sitemap) while keeping the pages
-// themselves fully served — nothing changes for their visitors.
+// (one shared image). If each self-asserted canonical over them, Google would
+// fold the duplicates into one winner per cluster and a public self-hosted
+// station could win a dispatch away from the official site. So non-official
+// installs donate those canonicals back and drop the pages from their sitemap,
+// while still serving them fully.
 export const IS_OFFICIAL_SITE =
   process.env.SUBWAVE_INDEX_ALL === '1' || isOfficialHost(SITE_URL);
